requirements.txt
fastapi
uvicorn
opencv-python
ultralytics
numpy
pillow
streamlit
requests
python-multipart